+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 9.21
+ - Fixed non-closing windows on some DoS pursuit mode attacks
+ - Added retrocompatibility for mdk3, added option in .airgeddonrc file to handle it
  - Added custom certificates creation for enterprise attacks
 
 ### 9.20